runelite-mixins: fix setting menu preventing menu add event from being fired
